% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.rq.prefixlist.df.R
\name{Get.rq.prefixlist.df}
\alias{Get.rq.prefixlist.df}
\title{Create prefix definition for SPARQL query from a prefix list}
\usage{
Get.rq.prefixlist.df(prefixlist)
}
\arguments{
\item{prefixlist}{List of prefixes made by qb.def.prefixlist}
}
\value{
String containing prefix statements for SPARQL query
}
\description{
The function defines the domain specifif prefixes used for the rdf
data cube for the Clinical Research and Development Analysis
Results Data
}

