% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.rdf.R
\name{update.rdf}
\alias{update.rdf}
\title{SPARQL update}
\usage{
update.rdf(model, sparqlUpdateStmts)
}
\arguments{
\item{model}{RRDF store}

\item{sparqlUpdateStmts}{SPARQL update statements string}
}
\value{
TRUE
}
\description{
SPARQL update
}

